各種画面を実装

+++ lib/acts_as_comable_customer/base.rb
@@ -4,20 +4,29 @@ module Comable::ActsAsComableCustomer
       base.extend ClassMethods
     end
 
-    class Comable::InvalidOrder < StandardError; end
-
     module ClassMethods
       def acts_as_comable_customer
         has_many :comable_orders, class_name: 'Comable::Order'
         alias_method :orders, :comable_orders
+
+        after_initialize :alias_methods_to_comable_customer_accsesor
+
         include InstanceMethods
       end
     end
 
     module InstanceMethods
-      def initialize(attributes={})
-        super
-        alias_methods_to_comable_customer_accsesor
+      require 'comable/customer_session'
+      require 'comable/cash_register'
+
+      include Comable::CustomerSession
+
+      def logged_in?
+        not self.new_record?
+      end
+
+      def not_logged_in?
+        not logged_in?
       end
 
       def add_cart_item(obj)
@@ -41,10 +50,12 @@ module Comable::ActsAsComableCustomer
       end
 
       def reset_cart
+        return super unless self.logged_in?
         self.cart_items.destroy_all
       end
 
       def cart_items
+        return super unless self.logged_in?
         customer_id = "#{Comable::Engine::config.customer_table.to_s.singularize}_id"
         Comable::CartItem.where(customer_id => self.id)
       end
@@ -57,42 +68,50 @@ module Comable::ActsAsComableCustomer
         def price
           self.sum(&:price)
         end
+
+        def products
+          self.map(&:product)
+        end
       end
 
-      def order(params={})
-        order = self.orders.build(params)
-        order.order_deliveries.map(&:order_details).flatten.each do |order_detail|
-          next unless order_detail.product
-          raise Comable::InvalidOrder unless self.remove_cart_item(order_detail.product)
-        end
-        order.save
+      def preorder(order_params={})
+        Comable::CashRegister.new(customer: self, order_attributes: order_params).build_order
+      end
+
+      def order(order_params={})
+        Comable::CashRegister.new(customer: self, order_attributes: order_params).create_order
       end
 
       private
 
       def add_product_to_cart(product)
+        return super unless self.logged_in?
+
         cart_items = find_cart_items_by(product)
         if cart_items.any?
           cart_item = cart_items.first
-          cart_item.update_attributes(quantity: cart_item.quantity.next)
+          cart_item.increment!(:quantity)
         else
           cart_items.create
         end
       end
 
       def remove_product_from_cart(product)
+        return super unless self.logged_in?
+
         cart_item = find_cart_items_by(product).first
         return false unless cart_item
 
-        cart_item.quantity = cart_item.quantity.pred
-        if cart_item.quantity.nonzero?
-          cart_item.save
+        if cart_item.quantity.pred.nonzero?
+          cart_item.decrement!(:quantity)
         else
           cart_item.destroy
         end
       end
 
       def find_cart_items_by(product)
+        return super unless self.logged_in?
+
         raise unless product.is_a?(Comable::Engine::config.product_table.to_s.classify.constantize)
 
         customer_id = "#{Comable::Engine::config.customer_table.to_s.singularize}_id"
